# optimized brigand::flatten