```cpp
using list = brigand::flatten<brigand::filled_list<int, SIZE>>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 35M | 00.05s - 36M | 00.07s - 39M | 00.12s - 49M | ko
gcc | 00.04s - 19M | 00.04s - 21M | 00.06s - 25M | 00.12s - 40M | 00.30s - 92M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 35M | 00.05s - 35M | 00.05s - 36M | 00.06s - 37M | 00.09s - 41M
gcc | 00.03s - 18M | 00.04s - 18M | 00.04s - 19M | 00.04s - 20M | 00.06s - 24M

---

```cpp
// list<list<list<...>, int>, int>
using list = brigand::flatten<
  brigand::fold<
    brigand::filled_list<int, SIZE>,
    brigand::list<>,
    brigand::list<brigand::_state, brigand::_element>
  >
>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.10s - 39M | 00.14s - 43M | 00.24s - 54M | ko
gcc | 00.05s - 21M | 00.09s - 28M | 00.13s - 37M | 00.23s - 57M | 00.46s - 106M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.09s - 39M | 00.14s - 43M | 00.23s - 53M | ko
gcc | 00.05s - 20M | 00.08s - 26M | 00.13s - 34M | 00.22s - 52M | 00.43s - 95M

---

```cpp
template<class N> struct rng { using type = brigand::range<int, 0, N::value>; };

// list<list<list<...>, range<int, 0, SIZE-1>>, range<int, 0, SIZE>>
using list = brigand::flatten<
  brigand::fold<
    brigand::range<int, 0, SIZE>,
    brigand::list<>,
    brigand::list<brigand::_state, rng<brigand::_element>>
  >
>;
```

Before:

Size= | 10 | 50 | 100 | 150
--- | --- | --- | --- | ---
clang | 00.08s - 37M | 00.37s - 71M | 01.57s - 238M | 04.25s - 643M
gcc | 00.07s - 24M | 00.43s - 103M | 02.12s - 420M | 05.46s - 1095M

After:

Size= | 10 | 50 | 100 | 150
--- | --- | --- | --- | ---
clang | 00.07s - 37M | 00.23s - 56M | 00.75s - 135M | 01.90s - 317M
gcc | 00.06s - 22M | 00.23s - 55M | 00.73s - 164M | 01.89s - 390M